requirements.txt
asgiref==3.6.0
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.4
django-filter==22.1
django-js-asset==2.0.0
django-mptt==0.14.0
djangorestframework==3.14.0
drf-yasg==1.21.4
exceptiongroup==1.0.4
Faker==15.3.4
idna==3.4
inflection==0.5.1
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
model-bakery==1.9.0
packaging==22.0
Pillow==9.3.0
pluggy==1.0.0
pytest==7.2.0
pytest-django==4.5.2
python-dateutil==2.8.2
pytz==2022.7
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
uritemplate==4.1.1
urllib3==1.26.13
